requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
textblob==0.18.0
numpy==1.26.4
twilio==9.2.3
pydantic==2.7.1
